Moved some code to other files
